requirements.txt
torch>=2.0.0
openai-whisper>=20231117
ffmpeg-python>=0.2.0